// src/app/modules/distributor/models/distributor.ts
export interface Distributor {
  code: string;
  description: string;
  distributorH1Index1Code: string;
  distributorH1Index2Code: string;
  distributorH1Index3Code: string;
  distributorH1Index4Code: string;
  distributorH1Index5Code: string;
  distributorH2Index1Code: string;
  distributorH2Index2Code: string;
  distributorH2Index3Code: string;
  distributorH2Index4Code: string;
  distributorH2Index5Code: string;
  main: boolean;
  active: boolean;
}